[应用]修复增加项目模块 [框架]增加新的模型形式,适用于修改记录,例如项目的状态记录.功能已完成,细节未优化,规范化 [资源]增加通用分类
git-svn-id: https://miku.googlecode.com/svn/trunk@115 a5e82c4f-c0ae-4e74-87f7-1291ebdd7ca5

+++ application/Trex/ActionController.php
@@ -226,6 +226,11 @@ class Trex_ActionController extends Trex_Controller
             $data = $meta->setForeignKeyData($meta['model'], $data);
 
             /**
+             * 保存关联模型的数据
+             */
+            $meta->saveRelatedDbData($meta, $data, $query);
+
+            /**
              * 入库
              */
             $ini = Qwin::run('-ini');
@@ -308,7 +313,12 @@ class Trex_ActionController extends Trex_Controller
              */
             $data = $meta->convertSingleData($relatedField, $relatedField, 'db', $_POST);
             $this->_meta->validateData($relatedField, $data + $_POST);
-            $data = $meta->restoreData($relatedField, $data);
+            $data = $meta->restoreData($editDbField, $data);
+
+            /**
+             * 保存关联模型的数据
+             */
+            $meta->saveRelatedDbData($meta, $data, $query);
 
             /**
              * 入库
@@ -348,11 +358,13 @@ class Trex_ActionController extends Trex_Controller
 
         // TODO $object->delete();
         // TODO 统计删除数
+        // TODO 删除数据关联的模块
         foreach($object as $key => $val)
         {
             foreach($meta['model'] as $model)
             {
-                if(isset($object[$key][$model['asName']]))
+                // 不删除视图模块的记录
+                if(isset($object[$key][$model['asName']]) && 'db' == $model['aim'])
                 {
                     $object[$key][$model['asName']]->delete();
                 }
